add create pod buttons to rich-text pod as well
