+++ ui/src/components/nodes/Code.tsx
@@ -59,7 +59,7 @@ import { useApolloClient } from "@apollo/client";
 import { NodeResizeControl, NodeResizer } from "reactflow";
 
 import "@reactflow/node-resizer/dist/style.css";
-import { ResizeIcon } from "./utils";
+import { NewPodButtons, ResizeIcon } from "./utils";
 
 export const ResultBlock = memo<any>(function ResultBlock({ id }) {
   const store = useContext(RepoContext)!;
@@ -356,163 +356,6 @@ function MyFloatingToolbar({ id, layout, setLayout }) {
         <DragIndicatorIcon fontSize="small" />
       </Box>
     </Box>
-  );
-}
-
-function NewPodButtons({ pod, xPos, yPos }) {
-  const store = useContext(RepoContext);
-  if (!store) throw new Error("Missing BearContext.Provider in the tree");
-  const addNode = useStore(store, (state) => state.addNode);
-  return (
-    <>
-      {/* Bottom 1 */}
-      <Button
-        variant="outlined"
-        size="small"
-        sx={{
-          // place it at the BOTTOM of the pod, centered
-          position: "absolute",
-          bottom: "0px",
-          left: "25%",
-          transform: "translate(-50%, 50%)",
-          zIndex: 100,
-          whiteSpace: "nowrap",
-          // opacity: showToolbar ? 0.5 : 0,
-          opacity: 0,
-          "&:hover": {
-            opacity: 1,
-          },
-        }}
-        onClick={() => {
-          addNode("CODE", { x: xPos, y: yPos + pod!.height! + 50 }, pod.parent);
-        }}
-      >
-        + Code
-      </Button>
-
-      {/* Bottom 2 */}
-      <Button
-        variant="outlined"
-        size="small"
-        sx={{
-          position: "absolute",
-          bottom: "0px",
-          left: "75%",
-          transform: "translate(-50%, 50%)",
-          zIndex: 100,
-          whiteSpace: "nowrap",
-          // opacity: showToolbar ? 0.5 : 0,
-          opacity: 0,
-          "&:hover": {
-            opacity: 1,
-          },
-        }}
-        onClick={() => {
-          addNode("RICH", { x: xPos, y: yPos + pod!.height! + 50 }, pod.parent);
-        }}
-      >
-        + Note
-      </Button>
-      {/* Left 1 */}
-      <Button
-        variant="outlined"
-        size="small"
-        sx={{
-          // place it at the LEFT of the pod, centered
-          position: "absolute",
-          top: "25%",
-          left: "0px",
-          transform: "translate(-50%, -50%)",
-          zIndex: 100,
-          whiteSpace: "nowrap",
-          // opacity: showToolbar ? 0.5 : 0,
-          opacity: 0,
-          "&:hover": {
-            opacity: 1,
-          },
-        }}
-        onClick={() => {
-          addNode("CODE", { x: xPos - pod!.width! - 50, y: yPos }, pod.parent);
-        }}
-      >
-        + Code
-      </Button>
-
-      {/* Left 2 */}
-      <Button
-        variant="outlined"
-        size="small"
-        sx={{
-          // place it at the LEFT of the pod, centered
-          position: "absolute",
-          top: "75%",
-          left: "0px",
-          transform: "translate(-50%, -50%)",
-          zIndex: 100,
-          whiteSpace: "nowrap",
-          // opacity: showToolbar ? 0.5 : 0,
-          opacity: 0,
-          "&:hover": {
-            opacity: 1,
-          },
-        }}
-        onClick={() => {
-          addNode("RICH", { x: xPos - pod!.width! - 50, y: yPos }, pod.parent);
-        }}
-      >
-        + Note
-      </Button>
-
-      {/* Right 1 */}
-      <Button
-        variant="outlined"
-        size="small"
-        sx={{
-          // place at the RIGHT of the pod, centered
-          position: "absolute",
-          top: "25%",
-          right: "0px",
-          transform: "translate(50%, -50%)",
-          zIndex: 100,
-          whiteSpace: "nowrap",
-          // opacity: showToolbar ? 0.5 : 0,
-          opacity: 0,
-          "&:hover": {
-            opacity: 1,
-          },
-        }}
-        onClick={() => {
-          addNode("CODE", { x: xPos + pod!.width! + 50, y: yPos }, pod.parent);
-        }}
-      >
-        + Code
-      </Button>
-
-      {/* Right 2 */}
-      <Button
-        variant="outlined"
-        size="small"
-        sx={{
-          // place at the RIGHT of the pod, centered
-          position: "absolute",
-          top: "75%",
-          right: "0px",
-          transform: "translate(50%, -50%)",
-          zIndex: 100,
-          whiteSpace: "nowrap",
-          // opacity: showToolbar ? 0.5 : 0,
-          opacity: 0,
-          "&:hover": {
-            opacity: 1,
-          },
-        }}
-        onClick={() => {
-          addNode("RICH", { x: xPos + pod!.width! + 50, y: yPos }, pod.parent);
-        }}
-      >
-        + Note
-      </Button>
-    </>
   );
 }
