+++ ui/src/components/nodes/utils.tsx
@@ -6,6 +6,14 @@ import {
   XYPosition,
 } from "reactflow";
 
+import { useContext } from "react";
+
+import Button from "@mui/material/Button";
+
+import { useStore } from "zustand";
+
+import { RepoContext } from "../../lib/store";
+
 export function ResizeIcon() {
   return (
     <svg
@@ -274,4 +282,161 @@ export function getHelperLines(
 
       return result;
     }, defaultResult);
+}
+
+export function NewPodButtons({ pod, xPos, yPos }) {
+  const store = useContext(RepoContext);
+  if (!store) throw new Error("Missing BearContext.Provider in the tree");
+  const addNode = useStore(store, (state) => state.addNode);
+  return (
+    <>
+      {/* Bottom 1 */}
+      <Button
+        variant="outlined"
+        size="small"
+        sx={{
+          // place it at the BOTTOM of the pod, centered
+          position: "absolute",
+          bottom: "0px",
+          left: "25%",
+          transform: "translate(-50%, 50%)",
+          zIndex: 100,
+          whiteSpace: "nowrap",
+          // opacity: showToolbar ? 0.5 : 0,
+          opacity: 0,
+          "&:hover": {
+            opacity: 1,
+          },
+        }}
+        onClick={() => {
+          addNode("CODE", { x: xPos, y: yPos + pod!.height! + 50 }, pod.parent);
+        }}
+      >
+        + Code
+      </Button>
+
+      {/* Bottom 2 */}
+      <Button
+        variant="outlined"
+        size="small"
+        sx={{
+          position: "absolute",
+          bottom: "0px",
+          left: "75%",
+          transform: "translate(-50%, 50%)",
+          zIndex: 100,
+          whiteSpace: "nowrap",
+          // opacity: showToolbar ? 0.5 : 0,
+          opacity: 0,
+          "&:hover": {
+            opacity: 1,
+          },
+        }}
+        onClick={() => {
+          addNode("RICH", { x: xPos, y: yPos + pod!.height! + 50 }, pod.parent);
+        }}
+      >
+        + Note
+      </Button>
+      {/* Left 1 */}
+      <Button
+        variant="outlined"
+        size="small"
+        sx={{
+          // place it at the LEFT of the pod, centered
+          position: "absolute",
+          top: "25%",
+          left: "0px",
+          transform: "translate(-50%, -50%)",
+          zIndex: 100,
+          whiteSpace: "nowrap",
+          // opacity: showToolbar ? 0.5 : 0,
+          opacity: 0,
+          "&:hover": {
+            opacity: 1,
+          },
+        }}
+        onClick={() => {
+          addNode("CODE", { x: xPos - pod!.width! - 50, y: yPos }, pod.parent);
+        }}
+      >
+        + Code
+      </Button>
+
+      {/* Left 2 */}
+      <Button
+        variant="outlined"
+        size="small"
+        sx={{
+          // place it at the LEFT of the pod, centered
+          position: "absolute",
+          top: "75%",
+          left: "0px",
+          transform: "translate(-50%, -50%)",
+          zIndex: 100,
+          whiteSpace: "nowrap",
+          // opacity: showToolbar ? 0.5 : 0,
+          opacity: 0,
+          "&:hover": {
+            opacity: 1,
+          },
+        }}
+        onClick={() => {
+          addNode("RICH", { x: xPos - pod!.width! - 50, y: yPos }, pod.parent);
+        }}
+      >
+        + Note
+      </Button>
+
+      {/* Right 1 */}
+      <Button
+        variant="outlined"
+        size="small"
+        sx={{
+          // place at the RIGHT of the pod, centered
+          position: "absolute",
+          top: "25%",
+          right: "0px",
+          transform: "translate(50%, -50%)",
+          zIndex: 100,
+          whiteSpace: "nowrap",
+          // opacity: showToolbar ? 0.5 : 0,
+          opacity: 0,
+          "&:hover": {
+            opacity: 1,
+          },
+        }}
+        onClick={() => {
+          addNode("CODE", { x: xPos + pod!.width! + 50, y: yPos }, pod.parent);
+        }}
+      >
+        + Code
+      </Button>
+
+      {/* Right 2 */}
+      <Button
+        variant="outlined"
+        size="small"
+        sx={{
+          // place at the RIGHT of the pod, centered
+          position: "absolute",
+          top: "75%",
+          right: "0px",
+          transform: "translate(50%, -50%)",
+          zIndex: 100,
+          whiteSpace: "nowrap",
+          // opacity: showToolbar ? 0.5 : 0,
+          opacity: 0,
+          "&:hover": {
+            opacity: 1,
+          },
+        }}
+        onClick={() => {
+          addNode("RICH", { x: xPos + pod!.width! + 50, y: yPos }, pod.parent);
+        }}
+      >
+        + Note
+      </Button>
+    </>
+  );
 }
